# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LibraryMCP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/books/": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "List Books",
        "description": "List all books with optional filters.\n\n- **title**: Filter books by title (Optional[str])\n- **author**: Filter books by author (Optional[str])\n- **genre**: Filter books by genre (Optional[str])\n- **available_only**: If True, only return books with available_copies \u003E 0 (bool)",
        "operationId": "list_books_books__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by book title (partial match)",
              "title": "Title"
            },
            "description": "Filter by book title (partial match)"
          },
          {
            "name": "author",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by author name (partial match)",
              "title": "Author"
            },
            "description": "Filter by author name (partial match)"
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by genre (partial match)",
              "title": "Genre"
            },
            "description": "Filter by genre (partial match)"
          },
          {
            "name": "available_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Show only books with available copies",
              "default": false,
              "title": "Available Only"
            },
            "description": "Show only books with available copies"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookResponse"
                  },
                  "title": "Response List Books Books  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Create Book",
        "description": "Add a new book to the catalog.\n\n- **book**: Book details for creation (BookCreate)\n\nReturns the created book with its assigned ID and available_copies set to total_copies.\nRaises 409 Error if a book with the same ISBN already exists.",
        "operationId": "create_book_books__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/books/{book_id}": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get Book",
        "description": "Retrieve a book by its ID.\n\n- **book_id**: The unique identifier of the book to retrieve (int)",
        "operationId": "get_book_books__book_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Books"
        ],
        "summary": "Update Book",
        "description": "Update an existing book in the database by book_id.\n\n- **book_id**: The unique identifier of the book to update (int)\n- **book_update**: The fields to update (BookUpdate)",
        "operationId": "update_book_books__book_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Book Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Books"
        ],
        "summary": "Delete Book",
        "description": "Delete a book from the database by book_id.\n\n- **book_id**: The unique identifier of the book to delete (int)",
        "operationId": "delete_book_books__book_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "examples": [
              "1984"
            ]
          },
          "author": {
            "type": "string",
            "title": "Author",
            "examples": [
              "George Orwell"
            ]
          },
          "isbn": {
            "type": "string",
            "title": "Isbn",
            "examples": [
              "978-0451524935"
            ]
          },
          "total_copies": {
            "type": "integer",
            "minimum": 1,
            "title": "Total Copies",
            "default": 1
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre",
            "examples": [
              "Fiction"
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "author",
          "isbn"
        ],
        "title": "BookCreate",
        "description": "Schema for creating a new Book."
      },
      "BookResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "examples": [
              "1984"
            ]
          },
          "author": {
            "type": "string",
            "title": "Author",
            "examples": [
              "George Orwell"
            ]
          },
          "isbn": {
            "type": "string",
            "title": "Isbn",
            "examples": [
              "978-0451524935"
            ]
          },
          "total_copies": {
            "type": "integer",
            "minimum": 1,
            "title": "Total Copies",
            "default": 1
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre",
            "examples": [
              "Fiction"
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "available_copies": {
            "type": "integer",
            "title": "Available Copies"
          }
        },
        "type": "object",
        "required": [
          "title",
          "author",
          "isbn",
          "id",
          "available_copies"
        ],
        "title": "BookResponse",
        "description": "Schema for Book responses."
      },
      "BookUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "isbn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn"
          },
          "total_copies": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Copies"
          },
          "available_copies": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Copies"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          }
        },
        "type": "object",
        "title": "BookUpdate",
        "description": "Schema for updating an existing Book."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}